Add endpoint for getting all the spell metadata names for later searching
The grouping should be handled by the DB query but CosmosDB doesn't appear to support
the group aggregation operation so the grouping needs to be done in the service instead.
There is still an issue of upper and lower case names being different but that is to be
solved separately.